# Context
- We recently mass deleted junk in authoring system including our old module object generator that doesn't work
- So we need to create a new module object generator/main that uses our newest parser code

# Change Summary
- Updated authoring system main
- Implemented module object generator + one test
- Made temporary changes to module.dart to get MOG to work, will still need to be fixed more soon

+++ luna_core/lib/renderers/divider_component_renderer.dart
@@ -37,7 +37,7 @@ class DividerComponentRenderer extends BaseComponentRenderer<DividerComponent> {
             painter: LinePainter(
               start: start,
               end: end,
-              thickness: component.thickness,
+              thickness: component.thickness.toDouble(),
               color: component.color,
             ),
             size: Size(constraints.maxWidth, constraints.maxHeight),
